Moved chunk support into standard API
