+++ pyfive/high_level.py
@@ -8,7 +8,7 @@ import posixpath
 import numpy as np
 
 from .core import Reference
-from .dataobjects import DataObjects
+from .dataobjects import DataObjects, DatasetDataObject
 from .misc_low_level import SuperBlock
 
 
@@ -89,7 +89,7 @@ class Group(Mapping):
         if dataobjs.is_dataset:
             if additional_obj != '.':
                 raise KeyError('%s is a dataset, not a group' % (obj_name))
-            return Dataset(obj_name, dataobjs, self)
+            return Dataset(obj_name, DatasetDataObject(self.file._fh, link_target), self)
         return Group(obj_name, dataobjs, self)[additional_obj]
 
     def __iter__(self):
